Update the source of Github actions
The source of all Github actions has been updated to point to frankhaugen/Workflows instead of the local actions. This allows for more consistency and prevents potential issues with local action discrepancies. Note that the migration includes not only the actions themselves, but also the configurations of how they are called in various workflow files.