Update method for fetching latest release version
This update changes the method in which the latest release version or tag is fetched in the action.yml file. Previously, the version was fetched manually with command line scripting and default input options. Now, it utilizes the existing 'oprypin/find-latest-tag' Github Action to fetch the version, simplifying the process and removing the need for default values.