Add GitHub actions for dotnet-script installation and execution
This commit includes the addition of new actions to install dotnet-script and another to run a C# script with .NET and Nuget package support. The workflow, which will be triggered on demand or upon a workflow call, is also updated to incorporate these new actions. This allows easy running of C# scripts in the context of this project.